requirements.txt
requests==2.27.1
python-dotenv==0.21.0

python-telegram-bot==11.1.0

vk_api==11.9.9
redis==4.5.1